Rename contact tracing stuff to household testing
